Rename modules and merge validation
