+++ library/__init__.py
@@ -1,0 +1,18 @@
+"""Utility library for chemical name normalization."""
+
+from .cleanup_competitor_io_validation import (
+    check_issues,
+    read_input_csv,
+    validate_input,
+    write_output_csv,
+)
+from .cleanup_competitor_names import normalize_name
+
+
+__all__ = [
+    "check_issues",
+    "normalize_name",
+    "read_input_csv",
+    "validate_input",
+    "write_output_csv",
+]
