--- mylib/__init__.py
@@ -1,15 +1,0 @@
-"""Utility library for chemical name normalization."""
-
-from .io_utils import read_input_csv, write_output_csv
-from .transforms import normalize_name
-
-from .validate import validate_input, check_issues
-
-
-__all__ = [
-    "read_input_csv",
-    "write_output_csv",
-    "normalize_name",
-    "validate_input",
-    "check_issues",
-]
